// tools/lib/file-formats/slices/types/dynamicslice.ts
export interface AZ__Component {
  __type: string
  id: number
}
// ...
export interface Asset {
  __type: string
  guid: string
  hint: string
  subId: number
  type: string
}
// ...
export interface DataFlagsPerEntity {
  __type: string
  entitytodataflags: []
}
// ...
export interface SliceComponent {
  __type: string
  baseclass1: AZ__Component
  dataflagsfornewentities: DataFlagsPerEntity
  dependencyreloadmode: number
  entities: Array<AZ__Entity>
  isdynamic: boolean
  prefabs: []
}